Renames images so that they're not blocked by add blockers
